refactor(GameResult): GameResult 도메인 객체로 분리
